Majority of pages working and navigatable.
All pages except for Feedback Dialogues, Edit Profile Dummy Data Filling, and Challenger Profile pages are working and manuverable, with the ability to accept challenges, where the challenge moves to the upcoming tab, opt out (reverse of accept), and swap between upcoming and completed feeds in the My Challenges tab. Each type of challenge (Main Feed, Accepted, and Completed) brings up the correct button on the challenge info page, and this page is auto-filled with the challenges data. Feeds refresh when returning to the previous page (from a secondary page) or navigating between pages using the navbar (from anywhere).